+++ CHANGELOG.md
@@ -11,11 +11,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - import_wudapt_fuel_model wrapper function to simplify import of WUDAPT urban fuel model
 - Sensitivity workflow scripts for rate of spread models using WUDAPT urban fuel model.
 
+### Fixed
+- Fuel moisture of extinction unit is `dimensionless` instead of `percent` in Anderson13.json
+
 ### Documentation
 - Add benchmark results for:
   - Rate of spread model sensitivity using WUDAPT urban fuel model:
     - Hamada 1
     - Hamada 2
+- Update benchmark results for:
+  - Rate of spread model sensitivity using Anderson 13 fuel model:
+    - Rothermel_SFIRE
 
 ## [0.3.1] - 2024 / 12 / 03
 ### Added